refactor: remove all mypy lenient overrides, fix 209 type errors
- Delete the 16-module override block from pyproject.toml
- Add type annotations across 62 files (no-untyped-def)
- Replace assert guards with explicit if+raise (production-safe)
- Fix Row/Any cascades with dict(), cast(), proper type narrowing
- Zero type: ignore added; mypy strict on all 140 source files

+++ nokori/posthoc/windowing.py
@@ -238,4 +238,4 @@ def _cosine_similarity(a: list[float], b: list[float]) -> float:
     nb = sum(y * y for y in b) ** 0.5
     if na == 0.0 or nb == 0.0:
         return 0.0
-    return dot / (na * nb)
+    return float(dot / (na * nb))
